move types to teacherFeedback directory
